requirements.txt
Pillow==10.3.0
Flask==3.0.3
